Solution to Check MSB of a Number

+++ Check MSB of a Number@@67882a731b4392d8fc1577e9/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main() {
-    long long int a;
-    scanf("%lld",&a);
+    int a;
+    scanf("%d",&a);
     if(a&1){
         printf("Set");
     }
